Refactor location for notice kinds functions
